fix(added alanda priorities)
