+++ frontend/alanda/apps/ui-demo/src/app/app.settings.ts
@@ -1,10 +1,4 @@
 import { SelectItem } from 'primeng/api';
-export const ALANDA_PRIORITIES: SelectItem[] = [
-  { value: -1, label: '-1 - Emergency plus' },
-  { value: 0, label: '0 - Emergency' },
-  { value: 1, label: '1 - Urgent' },
-  { value: 2, label: '2 - Normal' },
-];
 
 export const ALANDA_CONFIG: {
   LOCALE_PRIME: {
@@ -77,6 +71,5 @@ export const ALANDA_CONFIG: {
   AVATAR_BASE_PATH: 'src/app/assets',
   AVATAR_EXT: 'jpg',
   CLOSE_AFTER_COMPLETE: true,
-  PRIORITIES: ALANDA_PRIORITIES,
   /* WEBSOCKET_ENDPOINT : 'ws://localhost:8080/pmc-rest/websockets', */
 };
