+++ frontend/alanda/libs/common/src/lib/models/appSettings.ts
@@ -1,5 +1,6 @@
 import { InjectionToken, Injectable } from '@angular/core';
 import { LocaleSettings } from 'primeng/calendar';
+import { SelectItem } from 'primeng/api';
 
 @Injectable()
 export class AppSettings {
@@ -10,6 +11,14 @@ export class AppSettings {
   LOCALE_PRIME: LocaleSettings;
   APP_NAME?: string;
   CLOSE_AFTER_COMPLETE: boolean;
+  PRIORITIES: SelectItem[];
 }
+
+export const ALANDA_PRIORITIES: SelectItem[] = [
+  { value: -1, label: '-1 - Emergency plus' },
+  { value: 0, label: '0 - Emergency' },
+  { value: 1, label: '1 - Urgent' },
+  { value: 2, label: '2 - Normal' },
+];
 
 export const APP_CONFIG = new InjectionToken<AppSettings>('AppConfig');
